/* PR_START p--thread-local-id 01 */ Share `patch_id_generators` fixture.
