+++ metricflow-semantics/tests_metricflow_semantics/conftest.py
@@ -1,6 +1,7 @@
 # These imports are required to properly set up pytest fixtures.
 from __future__ import annotations
 
-from tests_metricflow_semantics.fixtures.id_fixtures import *  # noqa: F401, F403
+from metricflow_semantics.test_helpers.id_helpers import patch_id_generators  # noqa: F401
+
 from tests_metricflow_semantics.fixtures.manifest_fixtures import *  # noqa: F401, F403
 from tests_metricflow_semantics.fixtures.setup_fixtures import *  # noqa: F401, F403
