+++ tests_metricflow/conftest.py
@@ -1,9 +1,10 @@
 # These imports are required to properly set up pytest fixtures.
 from __future__ import annotations
 
+from metricflow_semantics.test_helpers.id_helpers import patch_id_generators  # noqa: F401
+
 from tests_metricflow.fixtures.cli_fixtures import *  # noqa: F401, F403
 from tests_metricflow.fixtures.dataflow_fixtures import *  # noqa: F401, F403
-from tests_metricflow.fixtures.id_fixtures import *  # noqa: F401, F403
 from tests_metricflow.fixtures.manifest_fixtures import *  # noqa: F401, F403
 from tests_metricflow.fixtures.setup_fixtures import *  # noqa: F401, F403
 from tests_metricflow.fixtures.sql_client_fixtures import *  # noqa: F401, F403
